Switch to async traits for bitcoind and KVStore outside of monitors
With LDK 0.2, most traits now offer async variants. In the previous
commit we ignored these but here we migrate to them everywhere
except in the `ChannelMonitor` persistence.

Note that we have to switch to async stdin reading in cli.rs to
ensure we always reach an `await` point in a timely manner on all
tasks.